Changes after BSD tests

+++ exrecover.c
@@ -4,7 +4,7 @@
 #include "ex_tty.h"
 
 #include <stdio.h>
-#include <sys/dir.h>
+#include <dirent.h>
 #include <paths.h>
 
 /*
@@ -567,7 +567,7 @@ null:
 				fprintf(stderr, " [Lost line(s):");
 			fprintf(stderr, " %d", was);
 			if ((ip - 1) - zero > was)
-				fprintf(stderr, "-%ld", (ip - 1) - zero);
+				fprintf(stderr, "-%d", (int)((ip - 1) - zero));
 			bad++;
 			was = 0;
 		}
@@ -576,7 +576,7 @@ null:
 			fprintf(stderr, " [Lost line(s):");
 		fprintf(stderr, " %d", was);
 		if (dol - zero != was)
-			fprintf(stderr, "-%ld", dol - zero);
+			fprintf(stderr, "-%d", (int)(dol - zero));
 		bad++;
 	}
 	if (bad)
